Revamp pup-python
Revamp `pup-python` by making the following larger changes:

1. We no longer install python2. We assume all systems only need python3
basis.
2. Add the ability to install pip packages into a specific virtualenv,
as well as globally.

These changes make the pup much more useful for FreeBSD, which doesn't
support conda.

Note... we assume that python3 and pip3 are called python and pip
respectively on the source system. We also assume that a virtualenv
executable is accessible. The first time the user uses this pup, they
may need to create a temporary virtualenv.