Update the Debian install component
We forgot to include `become: true` on the debian install component.
To be honest, I'm not even sure why this was triggering... but it should
be harmless. Also, just install python3 instead of python3 and
python3-pip. Long term, I need to decide if it even makes sense for
sheepdoge to attempt to manage python3 and python3-pip. While I'm
thinking about it, I want to limit the instances in which I download
certain software.